# internal/openapi/swagger.yaml
basePath: /api
definitions:
...
  api.APIErrorEnvelope:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      error:
        example: true
        type: boolean
      message:
        example: invalid request
        type: string
    type: object
...
  api.AddGeoFileRequest:
    properties:
      route:
        $ref: '#/definitions/api.DownloadRouteDTO'
      type:
        example: geosite
        type: string
      url:
        example: https://cdn.example.com/geosite.db
        type: string
    type: object
...
  api.GeoExpandData:
    properties:
      count:
        type: integer
      lines:
        items:
          type: string
        type: array
      path:
        type: string
    type: object
  api.GeoFileEntryDTO:
    properties:
      external:
        example: true
        type: boolean
      path:
        example: /opt/etc/hrneo/geosite.db
        type: string
      size:
        example: 3145728
        type: integer
      tagCount:
        example: 420
        type: integer
      type:
        example: geosite
        type: string
      updated:
        example: "2024-01-15T02:00:00Z"
        type: string
      url:
        example: https://cdn.example.com/geosite.db
        type: string
    type: object
  api.GeoFileResponse:
    properties:
      data:
        $ref: '#/definitions/api.GeoFileEntryDTO'
      success:
        example: true
        type: boolean
    type: object
...
  api.GeoFileUpdatedResponse:
    properties:
      data:
        $ref: '#/definitions/api.GeoFileUpdatedData'
      success:
        example: true
        type: boolean
    type: object
...
  api.GeoFilesRescannedResponse:
    properties:
      data:
        $ref: '#/definitions/api.GeoFilesRescannedData'
      success:
        example: true
        type: boolean
    type: object
  api.GeoFilesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.GeoFileEntryDTO'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
  api.HydraRouteConfigData:
    properties:
      autoStart:
        example: true
        type: boolean
      cidr:
        example: true
        type: boolean
      clearIPSet:
        example: false
        type: boolean
      conntrackFlush:
        example: true
        type: boolean
      directRouteEnabled:
        example: false
        type: boolean
      geoIPFiles:
        example:
        - /opt/etc/hrneo/geoip.db
        items:
          type: string
        type: array
      geoSiteFiles:
        example:
        - /opt/etc/hrneo/geosite.db
        items:
          type: string
        type: array
      globalRouting:
        example: false
        type: boolean
      ipsetEnableTimeout:
        example: false
        type: boolean
      ipsetMaxElem:
        example: 65536
        type: integer
      ipsetTimeout:
        example: 0
        type: integer
      log:
        example: warn
        type: string
      logFile:
        example: /var/log/hrneo.log
        type: string
      policyOrder:
        example:
        - default
        items:
          type: string
        type: array
    type: object
  api.HydraRouteConfigResponse:
    properties:
      data:
        $ref: '#/definitions/api.HydraRouteConfigData'
      success:
        example: true
        type: boolean
    type: object
...
  api.IpsetUsageResponse:
    properties:
      data:
        $ref: '#/definitions/api.IpsetUsageData'
      success:
        example: true
        type: boolean
    type: object
...
  api.OkResponse:
    properties:
      data:
        $ref: '#/definitions/api.OkData'
      success:
        example: true
        type: boolean
    type: object
...
  api.OversizedTagsResponse:
    properties:
      data:
        $ref: '#/definitions/api.OversizedTagsData'
      success:
        example: true
        type: boolean
    type: object
...
  api.PolicyOrderResponse:
    properties:
      data:
        $ref: '#/definitions/api.PolicyOrderData'
      success:
        example: true
        type: boolean
    type: object
...
  api.SetPolicyOrderRequest:
    properties:
      order:
        example:
        - default
        items:
          type: string
        type: array
    type: object
...
  api.TakeGeoFileControlRequest:
    properties:
      path:
        example: /opt/etc/HydraRoute/geosite_GA.dat
        type: string
    type: object
...
  api.UpdateGeoFileRequest:
    properties:
      path:
        example: /opt/etc/hrneo/geosite.db
        type: string
      route:
        $ref: '#/definitions/api.DownloadRouteDTO'
    type: object
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /hydraroute/config:
    get:
      description: Available when HydraRoute service is wired on the device.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HydraRouteConfigResponse'
      security:
      - CookieAuth: []
      summary: Get HydraRoute config
      tags:
      - hydraroute
  /hydraroute/config/update:
    put:
      consumes:
      - application/json
      description: Persists the HydraRoute (HrNeo) config and schedules a neo restart.
        The cached status becomes stale and is invalidated via SSE.
      parameters:
      - description: hydraroute.Config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.HydraRouteConfigData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HydraRouteConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Update HydraRoute config
      tags:
      - hydraroute
  /hydraroute/geo-expand:
    get:
      parameters:
      - description: geosite or geoip
        in: query
        name: kind
        required: true
        type: string
      - description: Tag name
        in: query
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GeoExpandData'
      security:
      - CookieAuth: []
      summary: HydraRoute geo tag expand
      tags:
      - hydraroute
  /hydraroute/geo-files:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GeoFilesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: List HydraRoute geo files
      tags:
      - hydraroute
  /hydraroute/geo-files/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Geo file source descriptor
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AddGeoFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GeoFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Add HydraRoute geo file
      tags:
      - hydraroute
  /hydraroute/geo-files/delete:
    delete:
      description: Removes the tracked geo data file at the given path and re-syncs
        the geo file list to config.
      parameters:
      - description: Filesystem path of the geo file
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Delete HydraRoute geo file
      tags:
      - hydraroute
  /hydraroute/geo-files/rescan:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GeoFilesRescannedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Rescan HydraRoute geo files from hrneo.conf
      tags:
      - hydraroute
  /hydraroute/geo-files/take-control:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filesystem path of the external geo file
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.TakeGeoFileControlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GeoFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Take HydraRoute geo file under awg-manager control
      tags:
      - hydraroute
  /hydraroute/geo-files/update:
    post:
      consumes:
      - application/json
      description: Empty path triggers a bulk refresh of every tracked geo file. Single
        path refreshes one file. Both branches return an updated count for caller-side
        feedback; the frontend refetches the list afterwards.
      parameters:
      - description: Path to refresh, or empty to refresh all
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateGeoFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GeoFileUpdatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Refresh HydraRoute geo file(s)
      tags:
      - hydraroute
  /hydraroute/geo-tags:
    get:
      parameters:
      - description: Geo file path
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - CookieAuth: []
      summary: HydraRoute geo tags
      tags:
      - hydraroute
  /hydraroute/ipset-usage:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.IpsetUsageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: HydraRoute ipset usage
      tags:
      - hydraroute
  /hydraroute/oversized-tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OversizedTagsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: HydraRoute oversized geo tags
      tags:
      - hydraroute
  /hydraroute/policy-order:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ordered list of policy names
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SetPolicyOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PolicyOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Set HydraRoute policy order
      tags:
      - hydraroute
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"